+++ src/providers/PortfolioProvider/types/positions.ts
@@ -1,6 +1,22 @@
-import type { DefiPosition } from '@/types/jumper-backend';
+import { orderBy, sumBy } from 'lodash';
+import type {
+  DefiPosition,
+  Chain,
+  Protocol,
+  Token,
+} from '@/types/jumper-backend';
 import type { PortfolioFormatters } from '../hooks/usePortfolioFormatters';
-import type { PortfolioExtendedToken } from './PortfolioExtendedToken';
+import { computePercentage } from '../utils/summary';
+import type { PortfolioExtendedToken, PriceLookup } from './tokens';
+
+export interface LpTokenIdentifier {
+  address: string;
+  chainId: number;
+}
+
+export type PositionGroupingFn = (position: PortfolioDefiPosition) => string;
+
+export type PositionGroupingKey = 'byProtocol' | 'byProtocolAndChain';
 
 interface PortfolioTokens {
   lpToken?: PortfolioExtendedToken;
@@ -70,5 +86,58 @@ export class PortfolioDefiPosition
 
   displayDebtUsd(options?: { compact?: boolean }): string {
     return this.displayAmountUSD(this.debtUsd, options);
+  }
+}
+
+export class PortfolioDeFiPositionsGroup {
+  main: PortfolioDefiPosition;
+  all: PortfolioDefiPosition[];
+  amountUSD: number;
+  percentageOfAmountUSD?: number;
+
+  private static formatters?: PortfolioFormatters;
+
+  static setFormatters(formatters: PortfolioFormatters): void {
+    PortfolioDeFiPositionsGroup.formatters = formatters;
+  }
+
+  constructor(
+    positions: PortfolioDefiPosition[],
+    totalPositionsValueUSD: number,
+  ) {
+    if (positions.length === 0) {
+      throw new Error(
+        'PortfolioDeFiPositionsGroup requires at least one position',
+      );
+    }
+
+    const sorted = orderBy(positions, (p) => p.netUsd, 'desc');
+    this.main = sorted[0];
+    this.all = sorted;
+    this.amountUSD = sumBy(sorted, (p) => p.netUsd);
+    this.percentageOfAmountUSD = computePercentage(
+      this.amountUSD,
+      totalPositionsValueUSD,
+    );
+  }
+
+  displayAmountUSD(): string {
+    if (PortfolioDeFiPositionsGroup.formatters) {
+      return PortfolioDeFiPositionsGroup.formatters.amountUSD(this.amountUSD);
+    }
+    return `$${this.amountUSD.toFixed(2)}`;
+  }
+
+  displayPercentageOfAmountUSD(): string {
+    if (PortfolioDeFiPositionsGroup.formatters && this.percentageOfAmountUSD) {
+      return PortfolioDeFiPositionsGroup.formatters.percentage(
+        this.percentageOfAmountUSD,
+      );
+    }
+    return `${(this.percentageOfAmountUSD ?? 0).toFixed(2)}%`;
+  }
+
+  hasMultiplePositions(): boolean {
+    return this.all.length > 1;
   }
 }
