+++ src/providers/PortfolioProvider/types/tokens.ts
@@ -1,3 +1,4 @@
+import { orderBy, sumBy } from 'lodash';
 import type { Token } from '@lifi/sdk';
 import type { ExtendedChain } from '@lifi/sdk';
 import { ExtendedToken } from '@/utils/Token';
@@ -5,17 +6,33 @@ import type { LiFiCommonToken } from '../lib/fetchTokensForAddresses';
 import type { Token as JumperToken, DefiToken } from '@/types/jumper-backend';
 import type { PortfolioFormatters } from '../hooks/usePortfolioFormatters';
 import { formatTokenAmount } from '@/utils/format';
+import { computePercentage } from '../utils/summary';
+
+export type PriceLookup = (
+  chainId: number,
+  address: string,
+) => number | undefined;
+
+export type GetTokenPrice = (
+  chainId: number,
+  address: string,
+) => number | undefined;
+
+export interface NormalizeTokensParams {
+  tokens: LiFiCommonToken[];
+  chains: ExtendedChain[];
+  getPrice: PriceLookup;
+}
+
+export type TokenGroupingFn = (token: PortfolioExtendedToken) => string;
+
+export type TokenGroupingKey = 'bySymbol' | 'byChain';
 
 interface PortfolioTokenData {
   amount: number;
   amountUSD: number;
   chainKey: string;
 }
-
-export type PriceLookup = (
-  chainId: number,
-  address: string,
-) => number | undefined;
 
 export class PortfolioExtendedToken extends ExtendedToken {
   amount: number;
@@ -141,5 +158,65 @@ export class PortfolioExtendedToken extends ExtendedToken {
       amountUSD,
       chainKey: token.chain.chainKey,
     });
+  }
+}
+
+export class PortfolioTokenGroup {
+  main: PortfolioExtendedToken;
+  all: PortfolioExtendedToken[];
+  amountUSD: number;
+  amount: number;
+  percentageOfAmountUSD?: number;
+
+  private static formatters?: PortfolioFormatters;
+
+  static setFormatters(formatters: PortfolioFormatters): void {
+    PortfolioTokenGroup.formatters = formatters;
+  }
+
+  constructor(tokens: PortfolioExtendedToken[], totalTokensValueUSD: number) {
+    if (tokens.length === 0) {
+      throw new Error('PortfolioTokenGroup requires at least one token');
+    }
+
+    const sorted = orderBy(tokens, (t) => t.amountUSD, 'desc');
+    this.main = sorted[0];
+    this.all = sorted;
+    this.amountUSD = sumBy(sorted, (t) => t.amountUSD);
+    this.amount = sumBy(sorted, (t) => t.amount);
+    this.percentageOfAmountUSD = computePercentage(
+      this.amountUSD,
+      totalTokensValueUSD,
+    );
+  }
+
+  displayAmountUSD(): string {
+    if (PortfolioTokenGroup.formatters) {
+      return PortfolioTokenGroup.formatters.amountUSD(this.amountUSD);
+    }
+    return `$${this.amountUSD.toFixed(2)}`;
+  }
+
+  displayAmount(): string {
+    if (PortfolioTokenGroup.formatters) {
+      return PortfolioTokenGroup.formatters.amount(
+        this.amount,
+        this.main.symbol,
+      );
+    }
+    return `${this.amount} ${this.main.symbol}`;
+  }
+
+  displayPercentageOfAmountUSD(): string {
+    if (PortfolioTokenGroup.formatters && this.percentageOfAmountUSD) {
+      return PortfolioTokenGroup.formatters.percentage(
+        this.percentageOfAmountUSD,
+      );
+    }
+    return `${(this.percentageOfAmountUSD ?? 0).toFixed(2)}%`;
+  }
+
+  hasMultipleChains(): boolean {
+    return this.all.length > 1;
   }
 }
