--- src/providers/PortfolioProvider/utils/computePercentage.ts
@@ -1,3 +1,0 @@
-export const computePercentage = (value: number, total: number): number => {
-  return total > 0 ? (value / total) * 100 : 0;
-};
