+++ src/providers/PortfolioProvider/utils/positions.ts
@@ -1,9 +1,34 @@
-import type { DefiPosition, DefiToken } from '@/types/jumper-backend';
-import { PortfolioDefiPosition } from '../types/PortfolioDefiPosition';
 import {
-  PortfolioExtendedToken,
-  type PriceLookup,
-} from '../types/PortfolioExtendedToken';
+  compact,
+  flatMap,
+  groupBy as groupByLodash,
+  orderBy,
+  values,
+} from 'lodash';
+import type { DefiPosition, DefiToken } from '@/types/jumper-backend';
+import {
+  PortfolioDefiPosition,
+  PortfolioDeFiPositionsGroup,
+  type LpTokenIdentifier,
+  type PositionGroupingFn,
+  type PositionGroupingKey,
+} from '../types/positions';
+import { PortfolioExtendedToken, type PriceLookup } from '../types/tokens';
+
+export const extractLpTokens = (
+  positions: DefiPosition[],
+): LpTokenIdentifier[] => {
+  return compact(
+    flatMap(positions, (position) =>
+      position.lpToken?.address && position.lpToken?.chain.chainId
+        ? {
+            address: position.lpToken.address,
+            chainId: position.lpToken.chain.chainId,
+          }
+        : null,
+    ),
+  );
+};
 
 const normalizeDefiToken = (
   token: DefiToken,
@@ -76,4 +101,36 @@ export const normalizePositions = (
   return positions
     .map((position) => normalizePosition(position, getPrice))
     .filter(Boolean) as PortfolioDefiPosition[];
+};
+
+export const groupByProtocol: PositionGroupingFn = (p) => p.protocol.name;
+
+export const groupByProtocolAndChain: PositionGroupingFn = (p) =>
+  `${p.protocol.name}-${p.chain.chainKey}`;
+
+export const positionGroupingFns: Record<
+  PositionGroupingKey,
+  PositionGroupingFn
+> = {
+  byProtocol: groupByProtocol,
+  byProtocolAndChain: groupByProtocolAndChain,
+};
+
+export const toPositionsGroup = (
+  positions: PortfolioDefiPosition[],
+): PortfolioDeFiPositionsGroup | null => {
+  if (positions.length === 0) {
+    return null;
+  }
+  return new PortfolioDeFiPositionsGroup(positions, 0);
+};
+
+export const groupPositions = (
+  positions: PortfolioDefiPosition[],
+  groupBy: PositionGroupingKey,
+): PortfolioDeFiPositionsGroup[] => {
+  const groupingFn = positionGroupingFns[groupBy];
+  const grouped = groupByLodash(positions, groupingFn);
+  const groups = values(grouped).map(toPositionsGroup);
+  return orderBy(compact(groups), (g) => g.amountUSD, 'desc');
 };
